Copy credential_process provider from babashka pod
I’m including this here as a discrete commit mainly to facilitate
better/easier peer review; this way a reviewer can diff the next commit
against this one to see exactly what changes I’ve made to the copy-pasta
code to adapt it to our needs.

tbh this isn’t the code *verbatim* but it’s close. I made a few tweaks
just so it would compile, e.g. I removed a reference to
`with-system-properties` and simplified function names a bit, minor
stuff like that.

CETUS-61