Adapt credential-process/provider to our needs
I have no idea why the original code in the pod is retrieving the config
key `credential_process` from the AWS CLI credentials file rather than
the config file.

In our case, wherein our profiles are created by aws-sso-util, that key
is in the profile definitions in our config files. So for this to work
for us, we need to retrieve this key from our config files.

I suppose perhaps there are some tools/people/cases that store that key
in the credentials file because I suppose the command could potentially
contain sensitive strings. I guess? But why would that matter? I mean,
it’s not like the credentials file is encrypted or something.

Regardless, it may be the case that the key *is* sometimes stored in the
credentials file, and we might want to support that case as well. TBD.

CETUS-61