Move CredentialsProviders into the same ns
I was just working on Qualia, updating it to use the credential_process
provider in addition to the aws-vault provider. I realized that the
developer ergonomics of having these in two discrete namespaces, using
two different naming conventions, was subpar.

This was an app that’s using both of these CredentialsProviders can do
so with a single entry in a `(:require)` stanza and with a consistent
naming convention.