improve ease of using Calva
…and possible various other Clojure IDEs? Not sure. Calva, at least,
doesn’t work well with deps.edn profiles that include :main-opts. So in
order to start a REPL that includes our test deps, I needed to remove
that key from the `:test` alias because otherwise any REPL I’d start
with that alias would run the tests and then exit.